+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 ### **Features**
 
 * `DynamicFormAutoFillService` added for importing and validating new HTML5 [autofill detail tokens](https://html.spec.whatwg.org/multipage/forms.html#autofill)
-
+* `ui-primeng` has arrived!
 
 # beta.2